refactor: search for cmds/events by extension
Refactored the command and event definitions to be found via their file
extension, `.command.js` and `.event.js` respectively. This allows me to
split off supporting code (datasets, class definitions, logic etc) into
seprate files for use elsewhere. For instance, I want my list of Dolores
ping quotes to be available outside the Ping command so that it can be
used for logging a welcome message to the console.