fix(maps): check next rotation is in bounds
I was getting a bug due to not checking whether the next index was
beyond the end of the array. I had already fixed this in the nextMap
getter, but ommitting this on the current map getter meant the app
crashed when it was at the end of the playlist.