refactor(embeds): move embeds to own directory
I will be using a lot of embeds for this project. At the moment they are
stored in directories with their triggers, but these are already
becoming split between commands and events. Storing them in one location
is clearer, because I will likely want to be able to create e.g.
Apex-related embeds from events, not just commands.