refactor(triggers): reduce closures usage
Moved a function in the Triggers definition into the class itself rather
than referencing a function declared outside. This creates a closure,
and although the performance hit for this is small, this may still
become an issue when receiving many messages. Also it's cleaner and
easier to read with the whole function declared as a method on Trigger -
it's a fairly minimal class anyway.