refactor(events): simplify registerEvents
Was previously following the discordjs recommended project structure,
reading event files programatically and dynamically loading them onto
the client.

I have decided to break away from this for three reasons. Firstly, there
aren't many events that I'll be monitoring, so this is unnecessarily
complex. Secondly, if I need to deactivate an event for bugfixing, it's
much easier to just quote out the unneeded event than to delete or move
the file somewhere - it's not like I'll be needing to live-reload them.
And thirdly, dynamically importing the events mandates using the
CommonJS format, which I'm planning on moving away from in favour of ES6
modules (MJS).