feat!: upgrade to ES6 module format
Moved from CommonJS to ES6 module (esm) format. Sacrifices some
flexibility with importing files (particularly JSON) and dynamic imports
in favour of a more pleasant and stricter syntax, potential tree-shaking
optimisations, and an easier upgrade route to TypeScript if I choose to
go down that route.

Forced a small rewrite of the deployCommands registrar to be more in
line with the others. This allowed me to drop the glob dependency.