test: add test coverage report
Fixes the `test` script and adds a coverage report option with
`test:coverage` using the `c8` package. This is a replacement for
Instanbul (aka `nyc`) and uses the same coverage report format, but
unlike `nyc` it works with native es6 modules (esm).

`c8` uses the native nodejs v8 engine to check for coverage. Because es6
modules are only loaded upon use, the `--all` flag is required to check
for whole-project coverage, as otherwise the engine will only load
modules for which tests already exist.

I decided to implement the coverage reports after reading a comment on
reddit from a hiring manager who says that coverage tests are a really
easy way to filter through applicants for coding jobs. Once I've put
this project on GitHub, I will be able to use github actions to apply a
badge to the readme that shows the current test coverage status.