feat(misquote): add misquote command
Misquote is a contextual command which transforms the text contents of
the message on which it was performed to a selected style, currently
`squidward` or `boomer`. Future transforms could include e.g. `zalgo` or
`furry`.

The transforms are already split off as utils to be used in future
commands, e.g. a slash-command version, or to use within phrase-based
responses.