feat!: rewrite as module
Changes to almost file in order to standardise formatting and rewrite
the project as a nodejs esm module for use in other projects. Not yet
complete with documentation, but complete enough that I should be able
to commit to a package repository for use in other personal projects.

BREAKING CHANGE: the entire API